fix(trace): decode unicode escapes in remaining JSON viewers
Follow-up to the PrettyJsonView fix (#13223). Extracts decodeUnicodeInJson
into a standalone util (avoiding a PrettyJsonView<->CodeJsonViewer import
cycle) and applies it across the remaining JSON display paths that still
rendered literal \uXXXX escapes:

- CodeJsonViewer (JSONView): decode parsed json in the useMemo, covering all
  its consumers (dataset runs, prompt detail, score metadata, evaluator
  preview, etc.)
- AdvancedJsonViewer via IOPreviewJSON: decode at the effective input/output/
  metadata source so search-match/comment offsets, rendering and copy stay
  consistent (decoding at the leaf would desync highlight offsets)
- Dataset item view/edit/diff (stringifyDatasetItemData)

Already-decoded strings are a no-op (decodeUnicodeEscapesOnly returns early
when there is no backslash).
