+++ web/src/features/datasets/utils/datasetItemUtils.ts
@@ -1,15 +1,19 @@
 import { showErrorToast } from "@/src/features/notifications/showErrorToast";
+import { decodeUnicodeEscapesOnly } from "@/src/utils/unicode";
 import type { Prisma } from "@langfuse/shared";
 
 /**
  * Converts a dataset item field value to a formatted JSON string.
  * Returns empty string for null/undefined values.
+ *
+ * \uXXXX escapes (e.g. Japanese ingested with Python ensure_ascii=True) are
+ * decoded so the value renders as real characters in the editor/viewer.
  */
 export const stringifyDatasetItemData = (data: unknown): string => {
   if (!data) return "";
 
   try {
-    return JSON.stringify(data, null, 2);
+    return decodeUnicodeEscapesOnly(JSON.stringify(data, null, 2), true);
   } catch {
     showErrorToast(
       "Failed to stringify data",
